Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I would be very unhappy with a setting for this. It is a minor detail the user should not care about. If we can autodetect and display the right bitmap, that's great, however if it needs manual setup, I don't think it's worth it.
> 
> I'm fine with whatever we display in the menu (as I mentionned on IRC, it will never fully match my keyboard due to the way we mix alt and windows keys to implement AltGr). But I'm not fine with a setting to choose what to display.

What we need is to detect Mac keyboards, and automatically switch the command and option keys when we detect one. This is the real bug. I believe that we can do this by looking at the usb vendor id and device id. USB vendor id 0x05AC is for Apple, and here is a list of all device ids: https://devicehunt.com/view/type/usb/vendor/05AC

Once this is done, we can worry about what bitmaps we use in the menus, and maybe switch to using key roles (control, option, command) instead of trying to map to the keys we think might be on your keyboard. At the current time, the current solution is the best we can do which is to assume a US PC keyboard and provide bitmaps that match the keys used there (ctrl, opt, and alt.) actually OPT is not good, it should be WIN or a symbol to represent the Windows key, but due to historical factors going  back to when BeOS ran on PPC Macs, and copyright of the Windows logo, this oddity has persisted to the current day.

Patch-set: 1
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>