Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > What we need is to detect Mac keyboards, and automatically switch the command and option keys when we detect one. This is the real bug. I believe that we can do this by looking at the usb vendor id and device id. USB vendor id 0x05AC is for Apple, and here is a list of all device ids: https://devicehunt.com/view/type/usb/vendor/05AC
> 
> I don't think this is currently possible. Normally you would use get_keyboard_id for this but it returns a single 16bit value and I don't remember what it gives for USB keyboards (I thought I had made it return the USB product ID but can't find that code anymore).

This may not be currently possible but we need to make it possible somehow for my idea to work. Why can’t we expose the USB vendor and device ids to the keyboard API?.

> But even then, what if you have multiple keyboards connected? For example you have a Mac laptop, but connected a PC keyboard to it, or the opposite?

I thought that the Input Preferences dialog allowed for multiple keyboard support now. We would switch command and option on the built in laptop keyboard, and not switch it for PC keyboard, by using the usb vendor and device ids.

> > Once this is done, we can worry about what bitmaps we use in the menus, and maybe switch to using key roles (control, option, command) instead of trying to map to the keys we think might be on your keyboard. At the current time, the current solution is the best we can do which is to assume a US PC keyboard and provide bitmaps that match the keys used there (ctrl, opt, and alt.) actually OPT is not good, it should be WIN or a symbol to represent the Windows key, but due to historical factors going  back to when BeOS ran on PPC Macs, and copyright of the Windows logo, this oddity has persisted to the current day.
> 
> Even on PC keyboards (at least the 104/105 keys ones) it does't match at all, and can't be made to match, because of the AltGr key which we handle as "option".

AltGr does not have any meaning on Haiku, that is a Windows thing. On Haiku both left and right option keys serve to input special characters. We have cleverly altered the international key maps to cover this fact, but there is no AltGr key in Haiku and there never will be.

> Here is a table of what currently happens:
> 
> Key          Menu label     Function
> Left CTRL    CTRL           Control
> Left Win     OPT            Option
> Left Alt     ALT            Command
> Right AltGr  OPT            Option
> Right Win    ALT            Command
> Right CTRL   CTRL           Control
> 
> So in this situation it's a good thing that our naming does not match any existing key on the keyboard, at least. As long as this remains (we are mapping 3 roles (control/option/command) into 4 types of keys (control/windows/alt/altgr)) there is no hope to use the actual key labels. If we want to use the actual key labels, first of all we need to fix that and introduce a new, separate keycode to differenciate AltGr. Otherwise, we have no option but stick to key roles (which is what the menus are doing currently, except they show "ALT" for the "Command" role, trying to hint users as to which key to press, but in fact making things more confusing, as on my European keyboard that role is actually reached using "left ALT, or right WIN key").

There is no AltGr. You are talking about the preference that was set in the key maps but this preference has no bearing on what Haiku is doing which is to treat the left and right keys be they Alt, Option, or Control identically.

There was never any hope to match the menus to the labels on your keyboard, the best we can hope for is to bless a keyboard and use that for everyone regardless of what they are using, which is what we are doing now or to abandon all pretense and use key roles, and hope that people will understand what they mean. ALT is not a role, it maps to the key we are guessing says “Alt” on your keyboard.

Patch-set: 1
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>