Update patch set 1
Patch Set 1:

> This may not be currently possible but we need to make it possible somehow for my idea to work. Why can’t we expose the USB vendor and device ids to the keyboard API?.

We can, but we must add new APIs for it, and it will be per-keyboard.

> 
> > But even then, what if you have multiple keyboards connected? For example you have a Mac laptop, but connected a PC keyboard to it, or the opposite?
> 
> I thought that the Input Preferences dialog allowed for multiple keyboard support now. We would switch command and option on the built in laptop keyboard, and not switch it for PC keyboard, by using the usb vendor and device ids.

Currently it does not allow for multiple keymaps, which would be needed for that (and for other reasons, too, my typical setup here is French AZERTY laptop with Spanish USB keyboard for example).

> AltGr does not have any meaning on Haiku, that is a Windows thing. On Haiku both left and right option keys serve to input special characters. We have cleverly altered the international key maps to cover this fact, but there is no AltGr key in Haiku and there never will be.

No, AltGr is an european keyboard thing, it is not a Windows thing. And we (well, BeOS) have made a nonsensical hack to support it somehow. It is not clever, it's ugly and broken. We need to fix it.

> There is no AltGr. You are talking about the preference that was set in the key maps but this preference has no bearing on what Haiku is doing which is to treat the left and right keys be they Alt, Option, or Control identically.

What I'm saying is, there should be an AltGr. There isn't, that's wrong, and that confuses everything. As long as this is not fixed, nothing will make sense for me as I have the left Alt and the right Windows key mapped to the same thing (that's the default setting for all keymaps which require an AltGr key to operate).

> There was never any hope to match the menus to the labels on your keyboard, the best we can hope for is to bless a keyboard and use that for everyone regardless of what they are using, which is what we are doing now or to abandon all pretense and use key roles, and hope that people will understand what they mean. ALT is not a role, it maps to the key we are guessing says “Alt” on your keyboard.

It maps to the left Alt *and* the right Windows key by default for all keymaps which have an AltGr key. So we are not able to "guess" anything here. We know that it's not possible to guess. We map it to two keys that have different labels on them. Let's fix that. Then we can talk about the other problems.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0