Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > This may not be currently possible but we need to make it possible somehow for my idea to work. Why can’t we expose the USB vendor and device ids to the keyboard API?.
> 
> We can, but we must add new APIs for it, and it will be per-keyboard.

So the same problem is present since the last time I looked. We need to extend the keyboard API to support multiple keyboards first, then we can introduce switching the keys for only Apple keyboards.

> > > But even then, what if you have multiple keyboards connected? For example you have a Mac laptop, but connected a PC keyboard to it, or the opposite?
> > 
> > I thought that the Input Preferences dialog allowed for multiple keyboard support now. We would switch command and option on the built in laptop keyboard, and not switch it for PC keyboard, by using the usb vendor and device ids.
> 
> Currently it does not allow for multiple keymaps, which would be needed for that (and for other reasons, too, my typical setup here is French AZERTY laptop with Spanish USB keyboard for example).

Multiple key maps are a different set of problems, which are related to this, but we don’t need to alter the key maps to support multiple keyboards, as we would be altering the behavior below the key map level, in input server or possibly in the keyboard driver, but more likely in input server.

> > AltGr does not have any meaning on Haiku, that is a Windows thing. On Haiku both left and right option keys serve to input special characters. We have cleverly altered the international key maps to cover this fact, but there is no AltGr key in Haiku and there never will be.
> 
> No, AltGr is an european keyboard thing, it is not a Windows thing. And we (well, BeOS) have made a nonsensical hack to support it somehow. It is not clever, it's ugly and broken. We need to fix it.

I looked into adding AltGr to the modifiers a while back and I found that, there is no reason to ever add AltGr support because there is a fundamental misunderstanding of the difference between the way BeOS, Haiku, and macOS work with regards to special characters vs. how Windows and Linux work. The option role maps to special characters, it maps to them on the left, it maps to them on the right, there’s no difference. That is how Macs work, and there is no AltGr key on any Mac keyboard anywhere in the world because macOS does not use it, and neither does Haiku. It’s not broken, you just have to understand that there is a difference from how Haiku and other systems input special characters.

> > There is no AltGr. You are talking about the preference that was set in the key maps but this preference has no bearing on what Haiku is doing which is to treat the left and right keys be they Alt, Option, or Control identically.
> 
> What I'm saying is, there should be an AltGr. There isn't, that's wrong, and that confuses everything. As long as this is not fixed, nothing will make sense for me as I have the left Alt and the right Windows key mapped to the same thing (that's the default setting for all keymaps which require an AltGr key to operate).

Nope there shouldn’t be, there is no point to ever adding support for it. It wouldn’t make any difference because you can already assign different roles to the left and right keys if you want to. The problem is with the keyboard itself not having the keys you want where you want them.

> > There was never any hope to match the menus to the labels on your keyboard, the best we can hope for is to bless a keyboard and use that for everyone regardless of what they are using, which is what we are doing now or to abandon all pretense and use key roles, and hope that people will understand what they mean. ALT is not a role, it maps to the key we are guessing says “Alt” on your keyboard.
> 
> It maps to the left Alt *and* the right Windows key by default for all keymaps which have an AltGr key. So we are not able to "guess" anything here. We know that it's not possible to guess. We map it to two keys that have different labels on them. Let's fix that. Then we can talk about the other problems.

We can’t ever fix that problem because we have no way to know what the key says on your keyboard, all we have is a number, and that number gives us no information about what the key is called or where it is on the keyboard, all we know is that you pushed key number 68. So since we can never do better than what we are currently doing I’m not sure how much of a point there is to changing it. We are doing the best that we possibly can do right now as far as identifying keys and putting a label in the menu that maybe matches what’s on your keyboard. The only way to do better is to read the USB vendor and device ids, and then make some decisions from there. If we could detect an Apple keyboard and switch the command and option keys then we could put the control, option, and command roles in the menu, and then we could tell people to buy an Apple keyboard. We could have a setting for the menu to switch between displaying roles or PC keyboard approximations, because otherwise we are just going to confuse people.

No it doesn’t. The key maps were altered to work this way, but by default the operating system treats both sides of the keyboard the same, so left and right control are the same, left and right alt are the same. You don’t typically have a right Windows key, the menu key is there instead, which is the issue for those that like to use the right side to input special characters. But there is still no AltGr key role in Haiku. Mac keyboards typically include a right option key so if you use that kind of keyboard, you’ll have an option key role on the right side to use, of course the option role will be assigned to the command key and the command role to the option key, because the buttons are in a different place physically than they are on a PC keyboard, and that’s the reason this patch set exists in the first place.

Patch-set: 1
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>