Update patch set 1
Patch Set 1:

> > > AltGr does not have any meaning on Haiku, that is a Windows thing. On Haiku both left and right option keys serve to input special characters. We have cleverly altered the international key maps to cover this fact, but there is no AltGr key in Haiku and there never will be.
> > 
> > No, AltGr is an european keyboard thing, it is not a Windows thing. And we (well, BeOS) have made a nonsensical hack to support it somehow. It is not clever, it's ugly and broken. We need to fix it.
> 
> I looked into adding AltGr to the modifiers a while back and I found that, there is no reason to ever add AltGr support because there is a fundamental misunderstanding of the difference between the way BeOS, Haiku, and macOS work with regards to special characters vs. how Windows and Linux work. The option role maps to special characters, it maps to them on the left, it maps to them on the right, there’s no difference. That is how Macs work, and there is no AltGr key on any Mac keyboard anywhere in the world because macOS does not use it, and neither does Haiku. It’s not broken, you just have to understand that there is a difference from how Haiku and other systems input special characters.

Well that's maybe how it works currently, but that's not how it should work. For me that's quite the definition of broken.

> Nope there shouldn’t be, there is no point to ever adding support for it. It wouldn’t make any difference because you can already assign different roles to the left and right keys if you want to. The problem is with the keyboard itself not having the keys you want where you want them.

I'm missing a role.

The keys I have on my keyboard are:
- Control
- Alt
- Alt Gr
- Windows

The roles are:
- Control
- Command
- Option

How can I make it more clear that one role is missing because I have 4 different type of keys, and only 3 roles to assign to them? This is wrong. Think what you want. My keyboard has an extra key role that yours doesn't have, and I want to use it. That BeOS never properly supported it is not my problem, let's fix that.

Don't try to explain me that "but Alt Gr and Windows key are the same thing". They are clearly not. We need to differenciate them. I want my window key to be used for stack and tile and other shortcuts. And I want my AltGr key to be used to type special characters meant to be typed this way. Anything else is nonsense. That's how my keyboard, and any keyboard in the EU, should work (except for Mac ones which goes a different way and require 3 finger combos instead). And do not tell me I should use a Mac keyboard. Their layout is completely different at least for the French ones (you can check the two keymaps in Haiku if you want), and I'm not going to re-learn typing or work with mislabelled keys all over.

> No it doesn’t. The key maps were altered to work this way, but by default the operating system treats both sides of the keyboard the same, so left and right control are the same, left and right alt are the same. 

They aren't with the default keymap for AltGr. The system does not trat "sides of the keyboard". It just gives keycodes to the keymap and the keymap outputs key roles. And currently our key roles have no logic to them and are all messed up because we are trying to map 4 different physical keys to 3 roles. This can't work.

> You don’t typically have a right Windows key, the menu key is there instead, which is the issue for those that like to use the right side to input special characters.

My keyboard of course has window keys on both sides like any standard 104/105 keys keyboard.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0