Update patch set 1
Patch Set 1:

> You think that you have an AltGr key on your keyboard because that’s what’s written on the key, but you don’t, it’s just a regular old Alt key, and Haiku doesn’t use Alt for special characters.

No. I have an alt key on the left and an AltGr key on the right. They should not do the same thing. The Alt Gr key should also not do the same thing as the Windows key. It's an extra key and currently Haiku does not handle it.

> I’m not telling you to use an Apple keyboard, I’m only trying to explain the differences between how Haiku and other operating systems handle modifier keys and how that affects the modifier key labels and locations for keyboards made for those operating systems.

There should be no differences. There is now, and that's what I'm saying is broken.

The Alt Gr key is an extra key and we should handle it as such.

> Unfortunately we can’t do any better than we are now by altering the key maps. I realize that you want to push the button that’s in a specific location on your keyboard to do a specific thing but you have to understand that no matter what you want, the key will always yield 0x5f. Other keyboards have a key that produces 0x5f in a different location so there is no way to fix this for you without breaking it for someone else except what we are currently doing which is to patch it in the keymap.

No, that's not a problem. The Alt Gr key has the same key code as the "right alt" key on keyboards that have that. It's easy to know that this key was pressed. This works already.

We need to handle this key specifically. Not "same as alt" or "same as windows". For this we need an extra modifier constant (B_ALTGR_KEY or whatever) because it's an extra key role. And we need keymaps to handle that.

> If we could do better, it would have to be through detecting the specific keyboard you are using like by looking at the usb device id.

No, you are not understanding what I'm saying.

> There is no need for a separate AltGr role because that role is already provided by the option role.

It is not. That's what I'm trying to tell you. On my keyboard/keymap I have both an AltGr key (for typing special characters) AND an option key (the Windows key). They are not the same thing. They are different things. They need different key roles. They can't use the same. Can I make this clearer?

> So if your keyboard has a Windows key and an Alt key on the right side, then what’s the problem? I guess where the keys are located is the problem, but that’s a preference, perhaps a commonly held one, but not a bug.

All EU keyboards are like this:

Control Windows Alt Space AltGr Windows Menu Control

There are:
- Two control keys
- Two windows keys
- One alt key
- One altgr key
- One menu key

> Haiku is working as intended which is to to use the Windows key to input special characters and Alt for shortcuts, unless you change that in the keymap.

That is not "working as intended". That's how you think it should work, except:
- It doesn't: as I said already, the left window and AltGr keys are used for special characters. The right window and left alt are used to enter shortcuts. With all our default keymaps that have an AltGr key, that's what we do.
- It's also not what I want it to do. I want the Windows key to be used for Stack&Tile and other things like that (maybe call it the "meta" role). I want the AltGr key to be used for special characters (let's call it the "option" role). I want the Alt key to be used for menu shortcuts (the "command" role). Currently Haiku cannot do this. We are missing a keyrole. This is not "as intended", it's a hack because someone tried to do something to handle keyboards with an AltGr key, without adding an extra key role for it. It is certainly not working as intended for anyone who uses an AltGr key (all European users).

> For the shortcuts used in menu, either we put the labels there that are on a standard US 105 key keyboard, meaning ctrl, the Windows logo, and alt and switch the labels according to the keymap, or we use roles, and we put control, option, and command in the menus, and we don’t change the labels based on your modifier settings in keymap.

Again, currently the mapping (by default as we ship Haiku) is as follows:

- The keys to trigger menu commands is left ALT and right Windows (they have different labels on them)
- The key to enter special characters is right AltGr and left Windows (they have different labels, too)

So in the current situation, we CANNOT use "the label that's on the key" because our mapping is nonsense. We should fix that. And to fix that, we need an extra key role to handle these keyboards correctly.

The AltGr key, physically, replaces right Alt. But, we use it as a kind-of Windows key. As a "compensation", we reassign the right windows key, to be right alt, leaving the left windows key, to be a Windows key. Blissfully ignoring how the keys are labelled, and making a nonsensical thing as a result.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0