Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > This boils down to a Mac vs. PC debate. Macs use option to input special characters, PCs use AltGr for that purpose. Haiku inputs special characters the Mac way, you want it to work the Windows way, so I guess we have to implement both ways, and we basically already have.
> 
> The "Mac way" in France is yet again a different keymap, which uses a lot of option+shift+something to reach special characters like | (option shift L) or ~ (option shift N). 
> 
> > I still don’t understand what makes AltGr different from right Alt that it needs its own role. Do you need a new column of special characters in the keymap for it, one that is different from option?
> 
> Currently we map AltGr to "option". We also map the left windows key to "option". We then use that key for two things:
> - Stack&Tile management
> - Special characters 
> 
> I expect that the left windows key is used for Stack&Tile
> I expect that the AltGr/right Alt key is used for special characters.
> 
> I don't want windows+e to print an € character. I don't need it to print anything, I would expect it to only do special shortcut (as in Windows: win+up=Zoom current window, win+down=minimize current window, and so on).
> 
> I also don't want my AltGr key to be able to trigger any of these shortcuts. There's no reason it would do that.
> 
> And also, I do not want my right Windows key to behave as a "Command" role like left Alt for no reason. Both my Windows keys should be usable the same way.
> 
> On keyboards/keymaps that don't have an AltGr key, yes, we should merge these two functions into the same key. I see that a bit similarly to how we handle 1-button or 2-button mouses (with long clicks or command+click or left+right click to reach functionnality normally only available using the 2nd and 3rd button). If we want this "degraded" mode to work, indeed we do not need a whole extra keymap layer. We have one key that allows access to the "option" keymap layer (AltGr in my case), and we have another, distinct key, that will be restricted to keyboard shortcuts, and get limited use by default (but users could then use it for other things: Meta or Super key in Terminal apps, custom shortcuts, ... if they decide to and their hardware allows it).

Here is a French Apple keyboard layout:

https://support.apple.com/library/content/dam/edam/applecare/images/en_US/keyboards/french_numeric.png

As you can see there is no AltGr key there. Special characters and inputted with option/alt which is to the right of the command key on the right side, the location which you find unacceptable. But that’s how special characters are inputted on macOS in France (and other places that use the French keyboard.)

Haiku is doing the same thing, except that we use Alt in place of command, and the Windows key in place of option/alt, i.e. we are using the keys that are in the same location as they are on a Mac keyboard, but on a PC keyboard. So there is no AltGr role, and there doesn’t need to be one.

Here is a French AZERTY keyboard layout: http://xahlee.info/kbd/ilayout/french_keyboard_layout_2017-04-30.svg

That one does have an AltGr key, because Windows uses AltGr to input special characters. But this has no meaning on Haiku because Haiku does NOT use AltGr to input special characters, it uses the Windows key. There is no AltGr on Haiku and there never will be because Haiku was designed to input special characters the way a Mac does with the keys flipped as compared to on an Apple keyboard so that they are in the same physical location on a PC keyboard.

Now you are going to tell me that the French Mac keyboard is different than the AZERTY one in some way, and while that is true, it doesn’t affect how Haiku inputs special characters... I don’t think.

The fact that the Command and Option roles are switched in the French keymap on Haiku is bad, it isn’t working as designed, it is trying to emulate how Windows inputs special characters rather than emulating the way macOS inputs special characters.

That said, I don’t mind that these key setting are flipped on the right side in the international keymaps on Haiku, but you have to understand that this is not how the system was designed, and that’s why it doesn’t work quite right. You’re suppose to input special characters with the right Windows key, and use right Alt for shortcuts, but for reasons that I can’t understand this is deemed unacceptable, even though that’s how it works on an Apple keyboard on macOS in France.

So tell me, what is the difference between the French Apple keyboard layout and the French PC keyboard layout other than AltGr? Why can’t we make Haiku’s French keymap identical to the one on the French Apple keyboard except with command and option flipped to match Alt/AltGr and the Windows key respectively? There must be some reason that I don’t understand that this wouldn’t work.

Patch-set: 1
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>