Update patch set 1
Patch Set 1:

> Here is a French Apple keyboard layout:

Yes thanks I know how my keyboards looks and operates.

> https://support.apple.com/library/content/dam/edam/applecare/images/en_US/keyboards/french_numeric.png
> 
> As you can see there is no AltGr key there. Special characters and inputted with option/alt which is to the right of the command key on the right side, the location which you find unacceptable. But that’s how special characters are inputted on macOS in France (and other places that use the French keyboard.)

As you can *not* see because Apple wants you to think this is nice and clean, a lot of characters are not printed on the keys. For example there is no | there. You need to use option + shift + L to type a pipe. There is no ~, you need to use option + shift + N to type it.

> Haiku is doing the same thing, except that we use Alt in place of command, and the Windows key in place of option/alt, i.e. we are using the keys that are in the same location as they are on a Mac keyboard, but on a PC keyboard. So there is no AltGr role, and there doesn’t need to be one.

Haiku is not doing the same thing.
- The keymap is based on the PC one (the special characters are in different positions)
- You DO use AltGr to type the special characters in Haiku. In our keymap it is an "option" key. You can also use left Windows which is also an option key. You cannot use right window which is a command key. This is because people really, really want to use AltGr to type special characters. They don't want to use the windows key. Because that's how the keyboard should operate and we are trying to follow the PC behavior, but we can't, because we're missing a key role (we have 3 key roles, and we need 4).

It is not a Mac vs PC thing. We are already trying to be on the PC side and failing at it. We need to add whatever is needed to actually use PC keyboards, not make excuses from our legacy to pretend we are trying to do things like Mac, especially when we clearly are not (again, the keymaps are set so that you use AltGr to type special characters, not the windows key, because that's what PC users expect, and not doing so would get people lost).
 
> Here is a French AZERTY keyboard layout: http://xahlee.info/kbd/ilayout/french_keyboard_layout_2017-04-30.svg
> 
> That one does have an AltGr key, because Windows uses AltGr to input special characters. But this has no meaning on Haiku because Haiku does NOT use AltGr to input special characters, it uses the Windows key. There is no AltGr on Haiku and there never will be because Haiku was designed to input special characters the way a Mac does with the keys flipped as compared to on an Apple keyboard so that they are in the same physical location on a PC keyboard.

Haiku maps the AltGr to option. Because we really need an AltGr key and that's the best we could do. But the result is confusing because then it maps the right Windows key to Command, and there is no reason for that.

> The fact that the Command and Option roles are switched in the French keymap on Haiku is bad, it isn’t working as designed, it is trying to emulate how Windows inputs special characters rather than emulating the way macOS inputs special characters.

It's doing what it can so the keyboard is usable. I don't care about macOS, I have a PC keyboard and I want to use my keyboard normally.

> 
> That said, I don’t mind that these key setting are flipped on the right side in the international keymaps on Haiku, but you have to understand that this is not how the system was designed, and that’s why it doesn’t work quite right. You’re suppose to input special characters with the right Windows key, and use right Alt for shortcuts, but for reasons that I can’t understand this is deemed unacceptable, even though that’s how it works on an Apple keyboard on macOS in France.

Of course it's unacceptable. My keyboard has an AltGr key and I want to use it as in other OS. Otherwise it's a completely different layout (which we also provide, btw, there is a "French (Mac)" keymap in Haiku for people with mac keyboards).

> 
> So tell me, what is the difference between the French Apple keyboard layout and the French PC keyboard layout other than AltGr? Why can’t we make Haiku’s French keymap identical to the one on the French Apple keyboard except with command and option flipped to match Alt/AltGr and the Windows key respectively? There must be some reason that I don’t understand that this wouldn’t work.

I mean, just look at your own pictures:
- On PC keyboards, 6 will type -, on Mac it will type a paragraph sign
- On PC keyboards, 8 will type _, on Mac it will type !
- On PC keyboards, € is AltGr + E, on Mac it is apparently option + $
- The key left to 1 is a "squared" symbol on PC, it is @/# on Mac
- The key next to right shift is !/paragraph sign on PC, =/+ on Mac
- $, *, pound, are in different positions (left of the enter key)
- Do I need to go on or can we agree that these are simply two completely different keymaps?
- As already mentionned, this nice picture of the Mac keyboard doesn't tell the whole truth: notice | and ~ and many other characters are completely missing, and you need option+shift+some letter to reach them. Not great for programmers, possibly OK for normal persons.

I don't know why, in France, Apple and IBM went with different keymaps. It seems to be mostly IBM's fault, as for example the Amiga and Atari keyboards would be closer to the Apple ones (but with some differences yet again). But that's how it is and we're not going to change that.

It seems you don't understand how important the AltGr key is, and how that makes it unthinkable to change its behavior, to the point that we had to hack in some kind of support in our keymaps, despite missing a key role for it. The key is needed to type characters like [] {} # ~ \ which you could argue are only for programmers, but also things like € and @ which everyone uses. Changing how you type them is going to confuse users, more than the current situation would. So that's what we have now: getting the AltGr key to work as expected for typing, at the cost of making the left Windows key also do that in addition to its role for Stack&Tile shortcuts, and making the right windows key behave as right command.

Oh an by the way, thought two keymaps were bad? Well, there is a third one which is a recently introduced standard: https://norme-azerty.fr/ we also support it in Haiku but for now it seems computer manufacturers are not using it. It spices things up further by introducing various characters reachable using AltGr + Shift (our normal PC-Azerty keyboard also has some of these, matching what Linux does, but with some improvements)

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0